monkey animation for home page
this took a while - but  learnt how to reference images and change them dynamically in the code file (in this case, with a dual timer). Maybe I could extrapolate this to an animation class so people can put monkeys that response to user input (for example, getting a question wrong), as in the original UI design there is a monkey in the bottom right corner.