Topic View Re-Organisation
I've implemented infrastructure to allow a topic to have two views - one for the quiz page, and one for the question page. The code should run as before, but your topic files have been renamed. For example,  BinaryAdditionTopicPageView has been renamed to BinaryAdditionQuizPageView.

If you want to create a new view, e.g. if you have implemented your quiz page, but want to start on your learn page, create the relevant .axaml and .axaml.cs files in the Views directory, and create the relevant .cs file in the ViewModels directory.

In order to link these views to the actual window, go into TopicLearnSelectorPageViewModel.cs and add the ViewModel you have created either to the learnTopics variable, or the quizTopics variable, depending on what you have implemented. When doing this, don't forget to increment the array size in the declaration.